Fix bug in arg_ReadValue
- Fixed bug where the nibble isn't converted to a uint64_t, leading to a false result for 64-bit values

+++ argparse.c
@@ -144,7 +144,7 @@ uint64_t arg_ReadValue(char* str) {
         if(nibble == -1) {
             return (uint64_t)-1;
         }
-        out += nibble << (idx * 4);
+        out += (uint64_t)nibble << (idx * 4);
         idx++;
         cur--;
     }
